refactor(modularization): Phase 5 step 5-S4a — turn_feedback_service move (task #26)
Move the smallest of the six conversation services into its canonical
domain home at ``aperag/domains/conversation/service/``. The legacy
path now re-exports ``TurnFeedbackService`` + the
``turn_feedback_service_global`` singleton so the existing
``views/chat.py`` caller keeps working without a rename sweep.

Scope-limited on purpose:

* ``turn_feedback_service`` has zero cross-service dependencies —
  only ``aperag.db.ops`` (DB abstraction, not a G1 ban target) and
  the domain's own Pydantic schemas from 5-S3. That makes it the
  cleanest first carve and validates the 5-S4 pattern before the
  heavier ``chat_service`` / ``bot_service`` moves land.
* The ``from aperag.schema import view_models`` import is rewritten
  to ``from aperag.domains.conversation.schemas import Feedback,
  TurnFeedback, TurnFeedbackList`` — the domain home cannot import
  the banned aggregate ``aperag.schema.view_models`` per G1.
* The remaining five services (``chat_service`` / ``bot_service`` /
  ``chat_collection_service`` / ``chat_document_service`` /
  ``chat_title_service``) land in 5-S4b / c / d as separate
  sub-commits. ``bot_service`` needs the QuotaOps DI wire-up (Q1
  ruling C); ``chat_title_service`` needs a plan for its
  ``default_model_service`` consumer-owned Protocol; ``chat_service``
  is the largest at 310 LOC.

Verify:
* ``from aperag.service.turn_feedback_service import TurnFeedbackService,
  turn_feedback_service_global`` OK via shim
* ``TurnFeedbackService is
  aperag.domains.conversation.service.turn_feedback_service.TurnFeedbackService``
  identity preserved
* ``pytest tests/unit_test/`` 671 / 29 skipped
* ``uvx ruff check`` + ``uvx ruff format --check`` clean
* G1: ``Grep "from aperag\.(db\.models|service|schema\.view_models)"
  aperag/domains/conversation/service/turn_feedback_service.py`` → 0
